4.0.0 - 1.20.6 support - 1.20.5/1.20.6 support - Better api documentation - Now using the Adventure API for Color via MiniMessage - Tested Folia support now fully working

+++ src/main/java/me/refracdevelopment/simplegems/api/events/impl/GemsPayEvent.java
@@ -7,11 +7,12 @@ import org.bukkit.entity.Player;
 @Getter
 public class GemsPayEvent extends SimpleEvent {
 
-    private final long gems;
+    private final double gems;
     private final Player target;
 
-    public GemsPayEvent(Player player, Player target, long amount) {
+    public GemsPayEvent(Player player, Player target, double amount) {
         super(player);
+
         this.target = target;
         this.gems = amount;
     }
